Refactor and document
