+++ src/utils/mod.rs
@@ -1,0 +1,3 @@
+//! A collection of utility functions for common operations.
+
+pub mod daemon;
